Refactor medicine and stats components, update asset paths, and add image picker dependency
