+++ src/app/(tabs)/stats.tsx
@@ -1,18 +1,17 @@
-import {StyleSheet, View} from "react-native";
-import {ThemedText} from "@/components/ThemedText";
+import { View } from "react-native";
+import {OmissionsComponent, PieComponent, StreakComponent, WeeklyProgressComponent} from "@/components/screens/stats";
 
-export default function Page() {
+export default function StatsPage() {
   return (
-      <View style={styles.container}>
-        <ThemedText>Stats Screen</ThemedText>
-      </View>
-  )
-}
+    <View className="w-full h-full flex flex-col gap-4 dark:border-2 dark:border-blue-300 mt-4">
+      <PieComponent />
 
-const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        justifyContent: "center",
-        alignItems: "center",
-    }
-});
+      <View className="mx-auto h-60 w-96 flex flex-row gap-4 justify-between rounded-xl">
+        <WeeklyProgressComponent />
+        <OmissionsComponent />
+      </View>
+
+      <StreakComponent />
+    </View>
+  );
+}
