+++ src/assets/stats/pieChart.png
