+++ src/components/screens/stats/StreakComponent.tsx
@@ -1,5 +1,5 @@
 import { ThemedText } from "@/components/ThemedText";
-import { View, Text } from "react-native";
+import { View } from "react-native";
 
 const StreakComponent = () => {
   return (
