+++ src/components/screens/stats/WeeklyProgressComponent.tsx
