+++ src/components/screens/stats/index.tsx
@@ -1,0 +1,4 @@
+export { default as OmissionsComponent } from './OmissionsComponent';
+export { default as PieComponent } from './PieComponent';
+export { default as StreakComponent } from './StreakComponent';
+export { default as WeeklyProgressComponent } from './WeeklyProgressComponent';
